[MOVE] 로고 이미지 삭제 관련 코드 leagueteam/infrastructure 로 이동 #163
