+++ src/main/java/com/sports/server/command/leagueteam/infrastructure/LogoImageEventHandler.java
@@ -1,6 +1,7 @@
-package com.sports.server.common.application;
+package com.sports.server.command.leagueteam.infrastructure;
 
-import com.sports.server.command.leagueteam.domain.LogoImageDeletedEvent;
+import com.sports.server.command.leagueteam.infrastructure.LogoImageDeletedEvent;
+import com.sports.server.common.application.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
